+++ test/Minimization/Metaheuristic/solvers/ant_colony.jl
@@ -1,0 +1,20 @@
+# Copyright 2025 Luis M. B. Varona
+#
+# Licensed under the MIT license <LICENSE or
+# http://opensource.org/licenses/MIT>. This file may not be copied, modified, or
+# distributed except according to those terms.
+
+"""
+    TestAntColony
+
+Test suite for ant colony optimization in the context of matrix bandwidth minimization.
+"""
+module TestAntColony
+
+using MatrixBandwidth
+using MatrixBandwidth.Minimization
+using Test
+
+# TODO: Write here
+
+end