+++ test/Minimization/Metaheuristic/solvers/psohc.jl
@@ -1,0 +1,21 @@
+# Copyright 2025 Luis M. B. Varona
+#
+# Licensed under the MIT license <LICENSE or
+# http://opensource.org/licenses/MIT>. This file may not be copied, modified, or
+# distributed except according to those terms.
+
+"""
+    TestPSOHC
+
+Test suite for particle swarm optimization with hill climbing (PSO-HC) in the context of
+matrix bandwidth minimization.
+"""
+module TestPSOHC
+
+using MatrixBandwidth
+using MatrixBandwidth.Minimization
+using Test
+
+# TODO: Write here
+
+end